# GODRIVER-2627 optimize latency selector