## Summary
Optimize allocations in latency server selector.

## Background & Motivation
Similar to #1107 i try to optimize allocations in latency server selector. Benchmark provided without changes:
```
BenchmarkLatencySelector
BenchmarkLatencySelector/AllFit
BenchmarkLatencySelector/AllFit-10         	   64196	     17866 ns/op	  111273 B/op	       9 allocs/op
BenchmarkLatencySelector/AllButOneFit
BenchmarkLatencySelector/AllButOneFit-10   	   71056	     17038 ns/op	  111273 B/op	       9 allocs/op
BenchmarkLatencySelector/HalfFit
BenchmarkLatencySelector/HalfFit-10        	  124551	      9748 ns/op	   53928 B/op	       8 allocs/op
BenchmarkLatencySelector/OneFit
BenchmarkLatencySelector/OneFit-10         	 1471053	       825.6 ns/op	    1320 B/op	       3 allocs/op
PASS
```

And with proposed optimization:
```
BenchmarkLatencySelector
BenchmarkLatencySelector/AllFit
BenchmarkLatencySelector/AllFit-10         	 1868062	       686.8 ns/op	     904 B/op	       2 allocs/op
BenchmarkLatencySelector/AllButOneFit
BenchmarkLatencySelector/AllButOneFit-10   	  155904	      7382 ns/op	   41864 B/op	       3 allocs/op
BenchmarkLatencySelector/HalfFit
BenchmarkLatencySelector/HalfFit-10        	  277989	      4409 ns/op	   21384 B/op	       3 allocs/op
BenchmarkLatencySelector/OneFit
BenchmarkLatencySelector/OneFit-10         	 1247785	       966.7 ns/op	    1800 B/op	       3 allocs/op
PASS
```